src/player/pyproject.toml
[project]
name = "player"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.1",
    "flet>=0.80.2",
    "httpx>=0.27.0",
    "pygame>=2.5.0",
    "python-dotenv>=1.2.1",
    "questionary>=2.1.1",
    "requests>=2.32.5",
    "typer>=0.20.0",
]
